+++ appTools/ToolSolderPaste.py
@@ -305,7 +305,18 @@ class SolderPaste(AppTool):
         self.ui_connect()
 
     def on_row_selection_change(self):
-        self.update_ui()
+        sel_model = self.ui.tools_table.selectionModel()
+        sel_indexes = sel_model.selectedIndexes()
+
+        # it will iterate over all indexes which means all items in all columns too but I'm interested only on rows
+        sel_rows = set()
+        for idx in sel_indexes:
+            sel_rows.add(idx.row())
+
+        # update UI only if only one row is selected otherwise having multiple rows selected will deform information
+        # for the rows other that the current one (first selected)
+        if len(sel_rows) == 1:
+            self.update_ui()
 
     def ui_connect(self):
         # on any change to the widgets that matter it will be called self.gui_form_to_storage which will save the